refactor(api): update the name of the docs schemas folder
