+++ apps/api/src/app/groups/docSchemas/index.ts
