+++ apps/api/src/app/groups/docSchemas/merkleProof.ts
